Move `pushrules` package to the `session` package
