+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/session/pushrules/RuleScope.kt
@@ -13,12 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.matrix.android.sdk.api.pushrules
+package org.matrix.android.sdk.api.session.pushrules
 
-import org.matrix.android.sdk.api.session.events.model.Event
-
-interface Condition {
-    fun isSatisfied(event: Event, conditionResolver: ConditionResolver): Boolean
-
-    fun technicalDescription(): String
+object RuleScope {
+    const val GLOBAL = "global"
 }
