+++ vector/src/main/java/im/vector/app/features/notifications/NotificationAction.kt
@@ -15,7 +15,7 @@
  */
 package im.vector.app.features.notifications
 
-import org.matrix.android.sdk.api.pushrules.Action
+import org.matrix.android.sdk.api.session.pushrules.Action
 
 data class NotificationAction(
         val shouldNotify: Boolean,
