+++ vector/src/main/java/im/vector/app/features/settings/notifications/StandardActions.kt
@@ -16,7 +16,7 @@
 
 package im.vector.app.features.settings.notifications
 
-import org.matrix.android.sdk.api.pushrules.Action
+import org.matrix.android.sdk.api.session.pushrules.Action
 
 sealed class StandardActions(
         val actions: List<Action>?
