+++ vector/src/main/java/im/vector/app/features/settings/notifications/VectorSettingsDefaultNotificationPreferenceFragment.kt
@@ -20,7 +20,7 @@ import android.os.Bundle
 import im.vector.app.R
 import im.vector.app.core.preference.VectorPreferenceCategory
 import im.vector.app.features.analytics.plan.MobileScreen
-import org.matrix.android.sdk.api.pushrules.RuleIds
+import org.matrix.android.sdk.api.session.pushrules.RuleIds
 
 class VectorSettingsDefaultNotificationPreferenceFragment :
     VectorSettingsPushRuleNotificationPreferenceFragment() {
