+++ vector/src/main/java/im/vector/app/features/settings/push/PushRulesViewModel.kt
@@ -23,7 +23,7 @@ import im.vector.app.core.di.SingletonEntryPoint
 import im.vector.app.core.platform.EmptyAction
 import im.vector.app.core.platform.EmptyViewEvents
 import im.vector.app.core.platform.VectorViewModel
-import org.matrix.android.sdk.api.pushrules.rest.PushRule
+import org.matrix.android.sdk.api.session.pushrules.rest.PushRule
 
 data class PushRulesViewState(
         val rules: List<PushRule> = emptyList()
